1. 修改内部函数位置; 2.修改BertWordPieceEncoder的部分功能
