+++ fastNLP/core/trainer.py
@@ -352,7 +352,8 @@ from .utils import _move_dict_value_to_device
 from .utils import _get_func_signature
 from .utils import _get_model_device
 from .utils import _move_model_to_device
-from .utils import _model_contains_inner_module
+from fastNLP.core._parallel_utils import _model_contains_inner_module
+
 
 class Trainer(object):
     """
