+++ fastNLP/embeddings/embedding.py
@@ -41,7 +41,12 @@ class Embedding(nn.Module):
         
         self.dropout = nn.Dropout(dropout)
         if not isinstance(self.embed, TokenEmbedding):
-            self._embed_size = self.embed.weight.size(1)
+            if hasattr(self, 'embed_size'):
+                self._embed_size = self.embed.embed_size
+            elif hasattr(self, 'embedding_dim'):
+                self._embed_size = self.embed.embedding_dim
+            else:
+                self._embed_size = self.embed.weight.size(1)
             if word_dropout>0 and not isinstance(unk_index, int):
                 raise ValueError("When drop word is set, you need to pass in the unk_index.")
         else:
